requirements.txt
aiohttp
discord.py
ecdsa
pycryptodome